'BaseSpec' moved to its own package
